Conflicts:
	IslandFurniture/IslandFurniture-ejb/src/java/IslandFurniture/StaticClasses/Helper/QueryMethods.java
	IslandFurnitureJamesTestClient/src/java/IslandFurniture/WAR/Manufacturing/ViewProductionPlanning.java
	IslandFurnitureJamesTestClient/web/productionplanning/index.xhtml

+++ IslandFurniture/IslandFurniture-ejb/src/java/IslandFurniture/EJB/SalesPlanning/SupplierManager.java
@@ -13,12 +13,15 @@ import IslandFurniture.EJB.Entities.ManufacturingFacility;
 import IslandFurniture.EJB.Entities.ProcuredStock;
 import IslandFurniture.EJB.Entities.ProcurementContract;
 import IslandFurniture.EJB.Entities.ProcurementContractDetail;
+import IslandFurniture.EJB.Entities.PurchaseOrder;
 import IslandFurniture.EJB.Entities.RetailItem;
 import IslandFurniture.EJB.Entities.Stock;
 import IslandFurniture.EJB.Entities.StockSupplied;
 import IslandFurniture.EJB.Entities.StockSuppliedPK;
 import IslandFurniture.EJB.Entities.Supplier;
 import static IslandFurniture.StaticClasses.Helper.QueryMethods.findCountryByName;
+import static IslandFurniture.StaticClasses.Helper.QueryMethods.findPCDByStockMFAndSupplier;
+import static IslandFurniture.StaticClasses.Helper.QueryMethods.findSupplierByName;
 import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.Stateful;
@@ -52,7 +55,7 @@ public class SupplierManager implements SupplierManagerLocal {
         supplier = (Supplier) em.find(Supplier.class, supplierId);
         return supplier;
     }
-    public Supplier addSupplier(String supplierName, String countryName) {
+    public Supplier addSupplier(String supplierName, String countryName, String phoneNo, String email) {
         Country country;
         Supplier supplier;
         ProcurementContract pc;
@@ -60,6 +63,12 @@ public class SupplierManager implements SupplierManagerLocal {
         try {
             System.out.println("SupplierManager.addSupplier()");
             country = findCountryByName(em, countryName);
+            supplier = findSupplierByName(em, supplierName);
+            if(supplier != null) {
+                System.out.println("Supplier " + supplierName + " already exists");
+                return supplier;
+            }
+            
             supplier = new Supplier();
             pc = new ProcurementContract();
             pcdList = new ArrayList<ProcurementContractDetail>();
@@ -70,6 +79,8 @@ public class SupplierManager implements SupplierManagerLocal {
             supplier.setName(supplierName);
             supplier.setCountry(country);
             supplier.setProcurementContract(pc);
+            supplier.setEmail(email);
+            supplier.setPhoneNumber(phoneNo);
             
             em.persist(supplier);
             return supplier;
@@ -78,7 +89,7 @@ public class SupplierManager implements SupplierManagerLocal {
             return null;
         }
     }
-    public void editSupplier(Long id, String name, String countryName) {
+    public void editSupplier(Long id, String name, String countryName, String phoneNumber, String email) {
         Supplier supplier;
         Country country;
         try {
@@ -89,24 +100,43 @@ public class SupplierManager implements SupplierManagerLocal {
                 supplier.setName(name);
             if(countryName != null) 
                 supplier.setCountry(country);
+            if(phoneNumber != null)
+                supplier.setPhoneNumber(phoneNumber);
+            if(email != null)
+                supplier.setEmail(email);
             em.persist(supplier);
         } catch(Exception ex) {
             System.out.println("Something went wrong");
         }
     }
     public void deleteSupplier(Long id) {
-        Supplier supplier;
+        Supplier supplier = null;
+        List<PurchaseOrder> poList;
         try {
             System.out.println("SupplierManager.deleteSupplier()");
-            supplier = em.find(Supplier.class, id);
-            System.out.println("Don't forget to mind the constraints. I'm gonna just delete for now");
-            for(int i=0; i<supplier.getProcurementContract().getProcurementContractDetails().size(); i++) 
-                em.remove(supplier.getProcurementContract().getProcurementContractDetails().get(i));
-            em.remove(supplier.getProcurementContract());
-            em.remove(supplier);
-            em.flush();
+            poList = em.createNamedQuery("getAllPurchaseOrders", PurchaseOrder.class).getResultList();
+            for(int i=0; i<poList.size(); i++) {
+                if(poList.get(i).getSupplier().getId().equals(id)) {
+                    supplier = poList.get(i).getSupplier();
+                    break;
+                }
+            }
+            
+            if(supplier != null) 
+                System.out.println("Existing purchase order linked to Supplier. Unable to delete");
+            else { 
+                supplier = em.find(Supplier.class, id);
+                for(int i=0; i<supplier.getProcurementContract().getProcurementContractDetails().size(); i++) 
+                    em.remove(supplier.getProcurementContract().getProcurementContractDetails().get(i));
+                System.out.println("Removed PCD"); 
+                em.remove(supplier.getProcurementContract());
+                System.out.println("Removed PC");
+                em.remove(supplier);
+                System.out.println("Removed supplier");
+                em.flush();
+            }
         } catch(Exception ex) {
-            System.err.println("Something went wrong");
+            System.err.println("Existing purchase order linked Supplier. Unable to delete");
         }
     }
     public List<ProcurementContractDetail> displayProcurementContractDetails(String supplierID) {
@@ -168,30 +198,35 @@ public class SupplierManager implements SupplierManagerLocal {
         List<ProcurementContractDetail> pcdList;
         ProcurementContractDetail pcd;
         try {
-            System.out.println("SupplierManager.addProcurementContractDetails()");
+            System.out.println("SupplierManager.addProcurementContractDetails()");            
             mf = em.find(ManufacturingFacility.class, mfID);
             stock = em.find(ProcuredStock.class, stockID);
             supplier = em.find(Supplier.class, supplierID);
             System.out.println("MF is " + mf.getName() + ", Stock is " + stock.getName() + ". Supplier is " + supplier.getName());
             
-            pc = supplier.getProcurementContract();
-            if(pc == null) {
-                pc = new ProcurementContract();
-                pcdList = new ArrayList<ProcurementContractDetail>();
-                pc.setProcurementContractDetails(pcdList);
-                pc.setSupplier(supplier);
-                supplier.setProcurementContract(pc);
+            pcd = findPCDByStockMFAndSupplier(em, stock, mf, supplier); 
+            
+            if(pcd != null) 
+                System.out.println("ProcurementContractDetail already exist");
+            else {            
+                pc = supplier.getProcurementContract();
+                if(pc == null) {
+                    pc = new ProcurementContract();
+                    pcdList = new ArrayList<ProcurementContractDetail>();
+                    pc.setProcurementContractDetails(pcdList);
+                    pc.setSupplier(supplier);
+                    supplier.setProcurementContract(pc);
+                }
+
+                pcd = new ProcurementContractDetail();
+                pcd.setProcuredStock(stock);
+                pcd.setSupplierFor(mf);
+                pcd.setProcurementContract(pc);
+                pcd.setLeadTimeInDays(leadTime);
+                pcd.setLotSize(size);
+
+                pc.getProcurementContractDetails().add(pcd);
             }
-            
-            pcd = new ProcurementContractDetail();
-            pcd.setProcuredStock(stock);
-            pcd.setSupplierFor(mf);
-            pcd.setProcurementContract(pc);
-            pcd.setLeadTimeInDays(leadTime);
-            pcd.setLotSize(size);
-            
-            pc.getProcurementContractDetails().add(pcd);
-            System.out.println("5");
         } catch(Exception ex) {
             ex.printStackTrace();
             System.err.println("Something went wrong here");
@@ -229,21 +264,16 @@ public class SupplierManager implements SupplierManagerLocal {
             System.out.println("SupplierManager.deleteStockSupplyRequest()");
             pk = new StockSuppliedPK(stockID, countryID, mfID);
             ss = em.find(StockSupplied.class, pk);
-            System.out.println("Found StockSupplied");
             if(ss == null)
                 System.out.println("StockSupplied is null");
-            mf = em.find(ManufacturingFacility.class, mfID);
-            System.out.println("1");
+            mf = em.find(ManufacturingFacility.class, mfID); 
             co = em.find(CountryOffice.class, countryID);
-            System.out.println("2");
-            
-            mf.getSupplyingWhatTo().remove(ss);
-            System.out.println("3"); 
-            co.getSuppliedWithFrom().remove(ss);
-            System.out.println("4");
             
             em.remove(ss);
-            System.out.println("5");
+            System.out.println("Successfully deleted Stock Supply Request");
+            mf.getSupplyingWhatTo().remove(ss);
+            co.getSuppliedWithFrom().remove(ss);
+            
             em.flush();
         } catch(Exception ex) {
             System.err.println("Something went wrong here");
@@ -254,22 +284,29 @@ public class SupplierManager implements SupplierManagerLocal {
         ManufacturingFacility mf;
         CountryOffice co;
         StockSupplied ss;
+        StockSuppliedPK pk;
         try {
             System.out.println("SupplierManager.addStockSupplyRequest()");
-            ss = new StockSupplied();
-            stock = em.find(Stock.class, stockID);
-            mf = em.find(ManufacturingFacility.class, mfID);
-            co = em.find(CountryOffice.class, countryID);
-            
-            ss.setCountryOffice(co);
-            ss.setManufacturingFacility(mf);
-            ss.setStock(stock);
-            
-            mf.getSupplyingWhatTo().add(ss);
-            co.getSuppliedWithFrom().add(ss);
-            
-            em.persist(ss);
-            em.flush();
+            pk = new StockSuppliedPK(stockID, countryID, mfID);
+            ss = em.find(StockSupplied.class, pk);
+            if(ss != null) 
+                System.out.println("Request already exists");
+            else {            
+                ss = new StockSupplied();
+                stock = em.find(Stock.class, stockID);
+                mf = em.find(ManufacturingFacility.class, mfID);
+                co = em.find(CountryOffice.class, countryID);
+
+                ss.setCountryOffice(co);
+                ss.setManufacturingFacility(mf);
+                ss.setStock(stock);
+
+                mf.getSupplyingWhatTo().add(ss);
+                co.getSuppliedWithFrom().add(ss);
+
+                em.persist(ss);
+                em.flush();
+            }
         } catch(Exception ex) {
             System.err.println("Something went wrong here");
         }
